@@ -1,7 +1,7 @@
 package com.moralesbatovski.pirateships.testing
 
 import android.annotation.SuppressLint
-import android.support.annotation.VisibleForTesting
+import androidx.annotation.VisibleForTesting
 import com.jakewharton.retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
 import okhttp3.HttpUrl
 import okhttp3.OkHttpClient
@@ -38,8 +38,7 @@ object DependencyProvider {
      */
     @SuppressLint("NewApi")
     fun getResponseFromJson(fileName: String): String {
-        val inputStream = javaClass.classLoader
-                .getResourceAsStream("$fileName.json")
+        val inputStream = javaClass.classLoader!!.getResourceAsStream("$fileName.json")
         val source = Okio.buffer(Okio.source(inputStream))
         return source.readString(StandardCharsets.UTF_8)
     }
