@@ -1,7 +1,7 @@
 package com.moralesbatovski.pirateships.mvvm.viewmodel
 
-import android.arch.lifecycle.ViewModel
-import android.arch.lifecycle.ViewModelProvider
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
 import io.reactivex.disposables.CompositeDisposable
 
 /**
